+++ src/main.ts
@@ -3,6 +3,7 @@ import { createPinia } from 'pinia'
 import router from './router'
 import i18n from './i18n'
 import App from './App.vue'
+import { tooltip } from './directives/tooltip'
 import './assets/main.css'
 
 const app = createApp(App)
@@ -10,5 +11,6 @@ const app = createApp(App)
 app.use(createPinia())
 app.use(router)
 app.use(i18n)
+app.directive('tooltip', tooltip)
 
 app.mount('#app')